Put all project related stuff in one module
