+++ src/components/accessWrappers/OrganizationEditRoute.tsx
@@ -1,6 +1,5 @@
 import Loading from '../elements/Loading';
-
-import NotFound from './NotFound';
+import NotFound from '../pages/NotFound';
 
 import useSafeParams from '@/hooks/useSafeParams';
 import { useHasOrganizationPermissions } from '@/modules/permissions/useHasPermissionsHooks';
