+++ src/components/clientDashboard/Service.tsx
@@ -3,8 +3,7 @@ import { useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import Loading from '../elements/Loading';
-
-import NotFound from './NotFound';
+import NotFound from '../pages/NotFound';
 
 import useSafeParams from '@/hooks/useSafeParams';
 import EditRecord from '@/modules/form/components/EditRecord';
