+++ src/components/pages/Organization.tsx
@@ -11,13 +11,13 @@ import TitleCard from '../elements/TitleCard';
 
 import NotFound from './NotFound';
 
+import OrganizationLayout from '@/components/layout/OrganizationLayout';
 import useSafeParams from '@/hooks/useSafeParams';
-import OrganizationDetails from '@/modules/inventory/components/OrganizationDetails';
-import OrganizationLayout from '@/modules/inventory/components/OrganizationLayout';
-import ProjectsTable from '@/modules/inventory/components/ProjectsTable';
-import { useOrganizationCrumbs } from '@/modules/inventory/components/useOrganizationCrumbs';
 import { OrganizationPermissionsFilter } from '@/modules/permissions/PermissionsFilters';
 import { useHasRootPermissions } from '@/modules/permissions/useHasPermissionsHooks';
+import OrganizationDetails from '@/modules/projects/components/OrganizationDetails';
+import ProjectsTable from '@/modules/projects/components/tables/ProjectsTable';
+import { useOrganizationCrumbs } from '@/modules/projects/hooks/useOrganizationCrumbs';
 import { Routes } from '@/routes/routes';
 import { PickListType, useDeleteOrganizationMutation } from '@/types/gqlTypes';
 import { evictPickList, evictQuery } from '@/utils/cacheUtil';
