+++ src/modules/projects/components/ProjectDetails.tsx
