+++ src/modules/projects/components/tables/InventoryTable.tsx
