+++ src/routes/routes.ts
@@ -4,15 +4,6 @@ export const Routes = {
   CLIENT_DASHBOARD: '/client/:clientId',
   ALL_PROJECTS: '/projects',
   PROJECT: '/projects/:projectId',
-  // EDIT_PROJECT: '/projects/:projectId/edit',
-  // NEW_INVENTORY: '/projects/:projectId/inventory/new',
-  // EDIT_INVENTORY: '/projects/:projectId/inventory/:inventoryId/edit',
-  // MANAGE_INVENTORY_BEDS: '/projects/:projectId/inventory/:inventoryId/beds',
-  // NEW_FUNDER: '/projects/:projectId/funder/new',
-  // EDIT_FUNDER: '/projects/:projectId/funder/:funderId/edit',
-  // NEW_COC: '/projects/:projectId/coc/new',
-  // EDIT_COC: '/projects/:projectId/coc/:cocId/edit',
-  // ADD_SERVICES: '/projects/:projectId/add-services',
   ORGANIZATION: '/organizations/:organizationId',
   EDIT_ORGANIZATION: '/organizations/:organizationId/edit',
   CREATE_PROJECT: '/organizations/:organizationId/new-project',
